+++ CHANGELOG.md
@@ -5,6 +5,8 @@
  
 - Added China-northwest, US-Gov-West & Paris regions
 
+- Allow setting both, Region name and endpoint, via `Region::Custom`
+
 ## [0.30.0] - 2017-12-02
 
 - Added CloudHSMv2